backend/requirements.txt
Django>=4.2.0,<4.3.0
djangorestframework>=3.14.0,<3.15.0
djangorestframework-simplejwt>=5.3.0,<5.4.0
psycopg2-binary>=2.9.9,<2.10.0
python-dotenv==1.0.1
drf-yasg>=1.21.7,<1.22.0
django-cors-headers>=4.3.1,<4.4.0
Pillow>=10.0.0,<10.1.0
django-storages==1.14.2
gunicorn==21.2.0
whitenoise>=6.6.0,<6.7.0
dj-database-url==2.1.0
reportlab==4.1.0
celery>=5.3.6,<5.4.0
redis>=5.0.1,<5.1.0
